Pointify `PDFLink`, adjust rectangle representation for annotations
`Rectangle` probably needs a more compact name.  Also,  I need to refactor the modules:  it would be nice to change the type of `annotationRect` from `a -> [PDFFloat]` to `a -> Rectangle`,  however `Rectangle` is defined in `Graphics.PDF.Shapes`, and `class AnnotationObject` is defined in `Graphics.PDF.Draw`,  and `Shapes` imports `Draw`,  causing a cyclic dependency.